Unilever/assignment2/requirements.txt
google-genai==1.12.1
langchain==0.3.24
langchain-community==0.3.23
langchain-google-genai==2.1.3
langchain-groq==0.3.2
# python-dotenv==1.1.0
requests==2.32.3